requirements.txt
SpeechRecognition==3.10.0
gTTS==2.3.1
PyAudio==0.2.13
openai==0.27.4
gTTS==2.3.1
python-vlc==3.0.18122